Run processing in lower priority so computer stays responding
