+++ geofileops/geofile.py
@@ -1194,7 +1194,6 @@ def _append_to_nolock(
             explodecollections=explodecollections,
             force_output_geometrytype=force_output_geometrytype,
             create_spatial_index=create_spatial_index,
-            priority_class='NORMAL',
             verbose=verbose)
     ogr_util.vector_translate_by_info(info=translate_info)
