+++ geofileops/util/geofileops_gpd.py
@@ -8,23 +8,23 @@ import datetime
 import enum
 import logging
 import logging.config
+import math
 import multiprocessing
 from pathlib import Path
-import pyproj 
 import time
 import shutil
-from typing import List, Optional, Tuple
+from typing import List, Optional, Tuple, NamedTuple
 import warnings
 
 # Don't show this geopandas warning...
 warnings.filterwarnings('ignore', 'GeoSeries.isna', UserWarning)
 
 import geopandas as gpd
+import psutil
 import shapely.geometry as sh_geom
 
 from geofileops import geofile
 from geofileops.util import general_util
-from geofileops.util.general_util import ParallelizationConfig
 from geofileops.util import geofileops_sql
 from geofileops.util import geometry_util
 from geofileops.util.geometry_util import GeometryType, PrimitiveType, SimplifyAlgorithm 
@@ -40,13 +40,112 @@ from geofileops.util import ogr_util
 logger = logging.getLogger(__name__)
 
 ################################################################################
-# The real work
+# Some helper functions
 ################################################################################
+
+class ParallelizationConfig():
+    def __init__(self, 
+            bytes_basefootprint: int = 50*1024*1024, 
+            bytes_per_row: int = 100, 
+            min_avg_rows_per_batch: int = 1000, 
+            max_avg_rows_per_batch: int = 10000, 
+            bytes_min_per_process = None, 
+            bytes_usable = None):
+        self.bytes_basefootprint = bytes_basefootprint
+        self.bytes_per_row = bytes_per_row
+        self.min_avg_rows_per_batch = min_avg_rows_per_batch
+        self.max_avg_rows_per_batch = max_avg_rows_per_batch
+        if bytes_min_per_process is None:
+            self.bytes_min_per_process = bytes_basefootprint + bytes_per_row * min_avg_rows_per_batch
+        else:
+            self.bytes_min_per_process = bytes_min_per_process
+        if bytes_usable is None: 
+            self.bytes_usable = psutil.virtual_memory().available * 0.9
+        else:
+            self.bytes_usable = bytes_usable
+
+parallelizationParams = NamedTuple('result', [('nb_parallel', int), ('nb_batches_recommended', int), ('nb_rows_per_batch', int)])
+def get_parallelization_params(
+        nb_rows_total: int,
+        nb_parallel: int = -1,
+        prev_nb_batches: int = None,
+        parallelization_config: ParallelizationConfig = None,
+        verbose: bool = False) -> parallelizationParams:
+    """
+    Determines recommended parallelization params.
+
+    Args:
+        nb_rows_total (int): The total number of rows that will be processed
+        nb_parallel (int, optional): The level of parallelization requested. 
+            If -1, tries to use all resources available. Defaults to -1.
+        prev_nb_batches (int, optional): If applicable, the number of batches 
+            used in a previous pass of the calculation. Defaults to None.
+        verbose (bool, optional): [description]. Defaults to False.
+
+    Returns:
+        parallelizationParams (NamedTuple('result', [('nb_parallel', int), ('nb_batches_recommended', int), ('nb_rows_per_batch', int)])): The recommended parameters.
+    """
+    # Init parallelization config
+
+    # If config is None, set to empty dict
+    if parallelization_config is not None:
+        parallelization_config_local = parallelization_config
+    else:
+        parallelization_config_local = ParallelizationConfig()
+    
+    # If the number of rows is really low, just use one batch
+    # TODO: for very complex features, possibly this limit is not a good idea
+    if nb_rows_total < parallelization_config_local.min_avg_rows_per_batch:
+        return parallelizationParams(1, 1, nb_rows_total)
+
+    if(nb_parallel == -1):
+        nb_parallel = multiprocessing.cpu_count()
+
+    logger.debug(f"memory_usable: {general_util.formatbytes(parallelization_config_local.bytes_usable)}, with:")
+    logger.debug(f"  -> mem.available: {general_util.formatbytes(psutil.virtual_memory().available)}")
+    logger.debug(f"  -> swap.free: {general_util.formatbytes(psutil.swap_memory().free)}") 
+
+    # If not enough memory for the amount of parallellism asked, reduce
+    if (nb_parallel * parallelization_config_local.bytes_min_per_process) > parallelization_config_local.bytes_usable:
+        nb_parallel = int(parallelization_config_local.bytes_usable/parallelization_config_local.bytes_min_per_process)
+        logger.debug(f"Nb_parallel reduced to {nb_parallel} to evade excessive memory usage")
+
+    # Optimal number of batches and rows per batch based on memory usage
+    nb_batches = math.ceil(
+            (nb_rows_total*parallelization_config_local.bytes_per_row*nb_parallel)/
+            (parallelization_config_local.bytes_usable-parallelization_config_local.bytes_basefootprint*nb_parallel))
+    
+    # Make sure the average batch doesn't contain > max_avg_rows_per_batch
+    batch_size = math.ceil(nb_rows_total/nb_batches)
+    if batch_size > parallelization_config_local.max_avg_rows_per_batch:
+        batch_size = parallelization_config_local.max_avg_rows_per_batch
+        nb_batches = math.ceil(nb_rows_total/batch_size)
+    mem_predicted = (parallelization_config_local.bytes_basefootprint + batch_size*parallelization_config_local.bytes_per_row)*nb_batches
+
+    # Make sure there are enough batches to use as much parallelism as possible
+    if nb_batches > 1 and nb_batches < nb_parallel:
+        if prev_nb_batches is None:
+            nb_batches = round(nb_parallel*1.25)
+        elif nb_batches < prev_nb_batches/4:
+            nb_batches = round(nb_parallel*1.25)
+    
+    batch_size = math.ceil(nb_rows_total/nb_batches)
+
+    # Log result
+    logger.debug(f"nb_batches_recommended: {nb_batches}, rows_per_batch: {batch_size}")
+    logger.debug(f" -> nb_rows_input_layer: {nb_rows_total}") 
+    logger.debug(f" -> will result in mem_predicted: {general_util.formatbytes(mem_predicted)}")
+
+    return parallelizationParams(nb_parallel, nb_batches, batch_size)
 
 class GeoOperation(enum.Enum):
     SIMPLIFY = 'simplify'
     BUFFER = 'buffer'
     CONVEXHULL = 'convexhull'
+
+################################################################################
+# The real stuff
+################################################################################
 
 def buffer(
         input_path: Path,
@@ -214,7 +313,7 @@ def _apply_geooperation_to_layer(
             nb_parallel = multiprocessing.cpu_count()
         nb_batches = nb_parallel
         batch_size = int(nb_rows_total/nb_batches)
-        #nb_parallel, nb_batches, batch_size = general_util.get_parallellisation_params(
+        #nb_parallel, nb_batches, batch_size = get_parallellisation_params(
         #        nb_rows_total=nb_rows_total,
         #        nb_parallel=nb_parallel,
         #        verbose=verbose)
@@ -232,7 +331,9 @@ def _apply_geooperation_to_layer(
             raise Exception(f"Error determining min_rowid and max_rowid for {temp_path}, layer {input_layer}")
         '''
 
-        with futures.ProcessPoolExecutor(nb_parallel) as calculate_pool:
+        with futures.ProcessPoolExecutor(
+                max_workers=nb_parallel, 
+                initializer=general_util.initialize_worker()) as calculate_pool:
 
             # Prepare output filename
             output_tmp_path = tempdir / output_path.name
@@ -518,7 +619,7 @@ def dissolve(
                 
                 # Calculate the best number of parallel processes and batches for 
                 # the available resources
-                nb_parallel, nb_batches_recommended, _ = general_util.get_parallelization_params(
+                nb_parallel, nb_batches_recommended, _ = get_parallelization_params(
                         nb_rows_total=nb_rows_total,
                         nb_parallel=nb_parallel,
                         prev_nb_batches=prev_nb_batches,
@@ -676,7 +777,9 @@ def _dissolve_polygons_pass(
     start_time = datetime.datetime.now()
     input_layerinfo = geofile.get_layerinfo(input_path, input_layer)
     nb_rows_total = input_layerinfo.featurecount
-    with futures.ProcessPoolExecutor(nb_parallel) as calculate_pool:
+    with futures.ProcessPoolExecutor(
+            max_workers=nb_parallel, 
+            initializer=general_util.initialize_worker()) as calculate_pool:
 
         batches = {}    
         future_to_batch_id = {}
